fix(#1367): break provider_sdk <-> providers.base import cycle
Fifth wedge for #1367. The cycle was:

- ``pollypm.provider_sdk`` did a TYPE_CHECKING import of
  ``LaunchCommand`` from ``pollypm.providers.base`` (and quoted the
  return-type annotation on ``ProviderAdapterBase.build_launch_command``).
- ``pollypm.providers.base`` did a top-level import of
  ``ProviderUsageSnapshot`` / ``TranscriptSource`` from
  ``pollypm.provider_sdk``.

Move ``LaunchCommand`` (a leaf dataclass with no further deps) into
``provider_sdk`` so the SDK module no longer needs any reference to
``providers.base``. Re-export ``LaunchCommand`` from ``providers.base``
for backwards compatibility — every existing
``from pollypm.providers.base import LaunchCommand`` keeps working
(supervisor, runtimes, probe_runner, claude/codex adapters,
default_launch_planner).

Verified:
- ``pollypm.provider_sdk`` ImportFrom edges no longer reference
  ``pollypm.providers.base`` (the cyclic dependency is gone in the
  static import graph).
- ``provider_sdk.LaunchCommand is providers.base.LaunchCommand`` — the
  re-export is identity, not a shadow class.
